refactor(stage1): move author logic to woeai.publications.authors
Candidate A, slice 3. Moves student-first-author detection, corresponding-
author detection, and the Degree Thesis Data loaders out of the script:

- woeai/publications/authors.py: load_degree_thesis_data (lru_cache, resolves
  docs/data/degree-theses.json from repo root), student_record_variants,
  iter_student_records, iter_current_student_authors, load_student_author_names,
  student_first_author_display, mark_student_first_author,
  has_corresponding_author_flag, explicit_corresponding_author_tokens,
  creator_name_index, corresponding_author_display_names,
  rendered_author_variants, mark_corresponding_authors.
- woeai/publications/textutils.py: gains creator_display_name,
  creator_csl_display_name, normalize_author_name (pure helpers the author
  logic depends on).
- Script re-exports all of these so tests (self.updater.<name>) are unchanged.
  Drops now-unused functools/html imports.

Script shrinks another 199 lines. 54 tests green.
